+++ src/lib/custom_control.ts
@@ -1,0 +1,4 @@
+export enum CUSTOM_CONTROL_TYPES {
+    LINE_CONTROL = 'lineControl',
+    CHANGE_DIRECTION_CONTROL = 'changeDirectionControl'
+}
