+++ src/lib/custom_control_zone.ts
@@ -5,12 +5,6 @@ import { CustomLineGroup, prepareVirtualLine } from "./custom_line";
 import { DirectionType } from "./zones";
 
 // http://fabricjs.com/custom-control-render
-
-export enum CUSTOM_CONTROL_TYPES {
-    LINE_CONTROL = 'lineControl',
-    CHANGE_DIRECTION_CONTROL = 'changeDirectionControl'
-}
-
 const cornerSize = 32
 const lineIcon = "data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='1.2rem' height='1.2rem' viewBox='0 0 24 24'%3E %3Crect width='100%25' height='100%25' stroke='rgba(0, 0, 0, 1)' stroke-width='2px' fill='rgba(255, 255, 255, 0.3)' %2F%3E %3Cpath fill='red' stroke='%234b4b4b' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='M4 18a2 2 0 1 0 4 0a2 2 0 1 0-4 0M16 6a2 2 0 1 0 4 0a2 2 0 1 0-4 0M7.5 16.5l9-9'/%3E%3C/svg%3E";
 const lineImageElem = document.createElement('img');
@@ -90,66 +84,6 @@ export const lineControl = new fabric.Control({
     cursorStyle: 'pointer',
     mouseUpHandler: lineControlHandler,
     render: lineRenderControlHandler,
-    sizeX: cornerSize,
-    sizeY: cornerSize
-})
-
-
-const changeDirectionIcon = "data:image/svg+xml,%3Csvg width='1.2rem' height='1.2rem' viewBox='0 0 1474.3327 1935.947' fill='red' xmlns='http://www.w3.org/2000/svg'%3E%3Crect width='100%25' height='100%25' stroke='rgba(0, 0, 0, 1)' stroke-width='5' fill='rgba(255, 255, 255, 0.3)' /%3E%3Cpath d='M513.166 1863.974h-192q-13 0-22.5-9.5t-9.5-22.5v-1376h-192q-13 0-22.5-9.5t-9.5-22.5q0-12 10-24l320-319q9-9 22-9 14 0 23 9l320 320q9 9 9 23 0 13-9.5 22.5t-22.5 9.5h-192v1376q0 13-9.5 22.5t-22.5 9.5m544 0q-14 0-23-9l-320-320q-9-9-9-23 0-13 9.5-22.5t22.5-9.5h192v-1376q0-13 9.5-22.5t22.5-9.5h192q13 0 22.5 9.5t9.5 22.5v1376h192q14 0 23 9t9 23q0 12-10 24l-319 319q-9 9-23 9' style='opacity:1;fill:red;stroke:%23000;stroke-width:42;stroke-miterlimit:4;stroke-dasharray:none;stroke-opacity:1'/%3E%3C/svg%3E"
-const changeDirectionElem = document.createElement('img');
-changeDirectionElem.src = changeDirectionIcon;
-
-const changeDirectionControlHandler = (eventData: MouseEvent, transformData: fabric.Transform, x: number, y: number): boolean => {
-    const targetObject = transformData.target
-    if (!targetObject) {
-        console.error('Empty target. Event: change_direction_control. Transform data:', transformData)
-        return false
-    }
-    if (!(targetObject instanceof CustomLineGroup)) {
-        console.error('Unhandled type. Only CustomLineGroup on top of fabric.Group has been implemented. Event: change_direction_control. Transform data:', transformData)
-        return false
-    }
-    if (!targetObject.parentContour) {
-        console.error('Empty parent contour. Event: change_direction_control. Transform data:', transformData)
-        return false
-    }
-    targetObject.direction = targetObject.direction === DirectionType.LeftRightTopBottom ? DirectionType.RightLeftBottomTop : DirectionType.LeftRightTopBottom
-    // Source group has 4 objects: [segment, directionText, L1Text, L2Text]
-    const directionTextObject = targetObject.getObjects()[1]
-    if (!(directionTextObject instanceof fabric.IText)) {
-        console.error('Unhandled object. Should be fabric.IText at position #1 Event: change_direction_control. Transform data:', transformData)
-        return false
-    }
-    directionTextObject.set('text', DirectionType.toString(targetObject.direction))
-    targetObject.parentContour.fire('virtial_line:modified', { target: targetObject.parentContour })
-    targetObject.canvas?.renderAll() // Force call of render
-    return true
-}
-
-const changeDirectionRenderControlHandler = (ctx: CanvasRenderingContext2D, left: number, top: number, styleOverride: any, fabricObject: fabric.Object): void => {
-    if (!(fabricObject instanceof fabric.Group)) {
-        // Render this control for this type of object only
-        // It still "transparent" and clickable, but error will be occured during click
-        // @todo: probably should wrap fabric.Group with custom grouping
-        return
-    }
-    ctx.save()
-    ctx.translate(left, top)
-    if (fabricObject.angle) {
-        ctx.rotate(fabric.util.degreesToRadians(fabricObject.angle))
-    }
-    ctx.drawImage(changeDirectionElem, -cornerSize/2, -cornerSize/2, cornerSize, cornerSize);
-    ctx.restore()
-}
-
-export const changeDirectionControl = new fabric.Control({
-    x: 0.5,
-    y: -0.5,
-    offsetY: -16,
-    offsetX: 16,
-    cursorStyle: 'pointer',
-    mouseUpHandler: changeDirectionControlHandler,
-    render: changeDirectionRenderControlHandler,
     sizeX: cornerSize,
     sizeY: cornerSize
 })
